chore: rename files to standarize casing
