+++ examples/lead-generation/src/components/views/display-errors.tsx
